move mocks to separate package to make them available in both, unit- and integration tests.
move mocks to separate package to make them available in both, unit- and integration tests.
