+++ mocks/lib/apiEncointer_mock.dart
@@ -2,10 +2,9 @@ import 'package:encointer_wallet/service/substrateApi/encointer/apiEncointer.dar
 import 'package:encointer_wallet/store/app.dart';
 import 'package:encointer_wallet/store/encointer/types/attestation.dart';
 import 'package:encointer_wallet/store/encointer/types/claimOfAttendance.dart';
-import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 
-import 'data/mockEncointerData.dart';
+import './data/mockEncointerData.dart';
 
 class MockApiEncointer extends Mock implements ApiEncointer {}
