+++ mocks/pubspec.yaml
@@ -1,0 +1,10 @@
+name: encointer_mocks
+description: Mocks for encointer unit tests and integration tests. (Separate package to make them available in both directories).
+
+environment:
+  sdk: ">=2.2.0 <3.0.0"
+
+dependencies:
+  encointer_wallet:
+    path: "./../"
+  mockito: any
