+++ test/page-encointer/meetup/attestation/components/stateMachinePartyB_test.dart
@@ -5,9 +5,9 @@ import 'package:encointer_wallet/service/substrateApi/api.dart';
 import 'package:encointer_wallet/store/app.dart';
 import 'package:encointer_wallet/store/encointer/types/attestationState.dart';
 
-import '../../../../mocks/apiEncointer_mock.dart';
-import '../../../../mocks/data/MockAccountData.dart';
-import '../../../../mocks/data/mockEncointerData.dart';
+import 'package:encointer_mocks/apiEncointer_mock.dart';
+import 'package:encointer_mocks/data/mockEncointerData.dart';
+import 'package:encointer_mocks/data/MockAccountData.dart';
 import 'common.dart';
 
 void main() {
