+++ test/store/account_test.dart
@@ -1,8 +1,8 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:encointer_wallet/store/app.dart';
 
-import '../mocks/data/MockAccountData.dart';
-import '../mocks/localStorage_mock.dart';
+import 'package:encointer_mocks/data/MockAccountData.dart';
+import 'package:encointer_mocks/localStorage_mock.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
